Refactor Tuya spell tests
This moves spell-related Tuya tests from `test_tuya.py` to `test_tuya_spells.py`.
